requirements.txt
playwright
python-dotenv
numpy
pandas
scikit-learn
requests
gguf
RapidFuzz
google-generativeai
python-telegram-bot>=21.0
aiohttp
nest-asyncio